requirements.txt
python-dotenv==1.0.1
fastapi==0.115.0
fastapi-cli==0.0.5
langgraph==0.2.22
pytest==8.3.3
pytest-asyncio==0.24.0
langchain_openai==0.2.1
# dev environment only
langfuse==2.52.0
langchain==0.3.3